Move custom commands to its own module
