+++ threepseat/commands/__init__.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import threepseat.commands.custom
 import threepseat.commands.emote
 import threepseat.commands.general
 import threepseat.commands.mmr
